PR17

+++ intbuffet/src/main/java/com/js/intbuffetproject/service/impl/CartServiceImpl.java
@@ -3,12 +3,20 @@ package com.js.intbuffetproject.service.impl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.js.intbuffetproject.model.Cart;
+import com.js.intbuffetproject.dto.Cart;
 import com.js.intbuffetproject.model.Item;
 import com.js.intbuffetproject.model.Product;
 import com.js.intbuffetproject.service.CartService;
 import com.js.intbuffetproject.service.ProductService;
 
+
+/**
+ * Class CartServiceImpl contains business logic related to class Cart.
+ * 
+ * @author Maria Borovtsova
+ * 
+ * @version 1.1
+ */
 @Service
 public class CartServiceImpl implements CartService {
 
@@ -17,11 +25,11 @@ public class CartServiceImpl implements CartService {
 
 	@Override
 	public Cart addToCart(Cart cart, Long id) {
-		Integer quantity = 0;
+		int quantity = 0;
 		if (cart.getProductsInCart().containsKey(id)) {
 
 			Item item = cart.getProductsInCart().get(id);
-			quantity = (Integer) cart.getProductsInCart().get(id).getQuantity() + 1;
+			quantity = cart.getProductsInCart().get(id).getQuantity() + 1;
 			item.setQuantity(quantity);
 
 			cart.addProduct(id, item);
@@ -41,7 +49,7 @@ public class CartServiceImpl implements CartService {
 
 		if (cart.getProductsInCart().containsKey(id)) {
 			Item item = cart.getProductsInCart().get(id);
-			Integer quantity = (Integer) cart.getProductsInCart().get(id).getQuantity();
+			int quantity = cart.getProductsInCart().get(id).getQuantity();
 			if (quantity > 0) {
 				--quantity;
 				item.setQuantity(quantity);
